@@ -8,7 +8,7 @@ void main() {
 
   const MethodChannel channel = MethodChannel('image_gallery_saver');
   final List<MethodCall> log = <MethodCall>[];
-  bool response;
+  bool? response;
 
   channel.setMockMethodCallHandler((MethodCall methodCall) async {
     log.add(methodCall);
@@ -23,7 +23,7 @@ void main() {
   test('saveImageToGallery test', () async {
     response = true;
     Uint8List imageBytes = Uint8List(16);
-    final bool result = await ImageGallerySaver.saveImage(imageBytes);
+    final bool? result = await (ImageGallerySaver.saveImage(imageBytes) as FutureOr<bool?>);
     expect(
       log,
       <Matcher>[
